+++ openstack_plugin/resources/compute/image.py
@@ -23,7 +23,8 @@ from openstack_plugin.decorators import (with_openstack_resource,
 from openstack_plugin.constants import (RESOURCE_ID, IMAGE_OPENSTACK_TYPE)
 from openstack_plugin.utils import (validate_resource_quota,
                                     reset_dict_empty_keys,
-                                    add_resource_list_to_runtime_properties)
+                                    add_resource_list_to_runtime_properties,
+                                    cleanup_runtime_properties)
 
 
 @with_compat_node
@@ -45,10 +46,16 @@ def start(openstack_resource):
 
 
 @with_compat_node
-@with_openstack_resource(OpenstackImage, ignore_unexisted_resource=True)
+@with_openstack_resource(OpenstackImage)
 def delete(openstack_resource):
+    if not ctx.instance.runtime_properties.get(RESOURCE_ID):
+        ctx.logger.info('Image is already uninitialized.')
+        return
     # Delete the image resource after lookup the resource_id values
     openstack_resource.delete()
+    cleanup_runtime_properties(ctx, [
+        RESOURCE_ID
+    ])
 
 
 @with_compat_node
