+++ openstack_plugin/resources/compute/keypair.py
@@ -22,7 +22,8 @@ from openstack_plugin.decorators import (with_openstack_resource,
                                          with_compat_node)
 from openstack_plugin.constants import (RESOURCE_ID, KEYPAIR_OPENSTACK_TYPE)
 from openstack_plugin.utils import (validate_resource_quota,
-                                    add_resource_list_to_runtime_properties)
+                                    add_resource_list_to_runtime_properties,
+                                    cleanup_runtime_properties)
 
 
 @with_compat_node
@@ -42,13 +43,19 @@ def create(openstack_resource):
 
 
 @with_compat_node
-@with_openstack_resource(OpenstackKeyPair, ignore_unexisted_resource=True)
+@with_openstack_resource(OpenstackKeyPair)
 def delete(openstack_resource):
     """
     Delete current openstack keypair
     :param openstack_resource: instance of openstack keypair resource
     """
+    if not ctx.instance.runtime_properties.get(RESOURCE_ID):
+        ctx.logger.info('KeyPair is already uninitialized.')
+        return
     openstack_resource.delete()
+    cleanup_runtime_properties(ctx, [
+        RESOURCE_ID, 'private_key', 'public_key'
+    ])
 
 
 @with_compat_node
