+++ openstack_plugin/resources/identity/project.py
@@ -32,7 +32,8 @@ from openstack_plugin.constants import (RESOURCE_ID,
 
 from openstack_plugin.utils import (validate_resource_quota,
                                     reset_dict_empty_keys,
-                                    add_resource_list_to_runtime_properties)
+                                    add_resource_list_to_runtime_properties,
+                                    cleanup_runtime_properties)
 
 
 def _assign_users(project_resource, users):
@@ -162,13 +163,19 @@ def start(openstack_resource, quota_dict={}):
 
 
 @with_compat_node
-@with_openstack_resource(OpenstackProject, ignore_unexisted_resource=True)
+@with_openstack_resource(OpenstackProject)
 def delete(openstack_resource):
     """
     Delete current openstack project
     :param openstack_resource: instance of openstack project resource
     """
+    if not ctx.instance.runtime_properties.get(RESOURCE_ID):
+        ctx.logger.info('Project is already uninitialized.')
+        return
     openstack_resource.delete()
+    cleanup_runtime_properties(ctx, [
+        RESOURCE_ID
+    ])
 
 
 @with_compat_node
