+++ openstack_plugin/resources/network/network.py
@@ -25,7 +25,8 @@ from openstack_plugin.decorators import (with_openstack_resource,
 from openstack_plugin.constants import RESOURCE_ID
 from openstack_plugin.utils import (validate_resource_quota,
                                     reset_dict_empty_keys,
-                                    add_resource_list_to_runtime_properties)
+                                    add_resource_list_to_runtime_properties,
+                                    cleanup_runtime_properties)
 
 from openstack_plugin.constants import NETWORK_OPENSTACK_TYPE
 
@@ -57,13 +58,19 @@ def create(openstack_resource):
 
 
 @with_compat_node
-@with_openstack_resource(OpenstackNetwork, ignore_unexisted_resource=True)
+@with_openstack_resource(OpenstackNetwork)
 def delete(openstack_resource):
     """
     Delete current openstack network
     :param openstack_resource: instance of openstack network resource
     """
+    if not ctx.instance.runtime_properties.get(RESOURCE_ID):
+        ctx.logger.info('Network is already uninitialized.')
+        return
     openstack_resource.delete()
+    cleanup_runtime_properties(ctx, [
+        RESOURCE_ID
+    ])
 
 
 @with_compat_node
