+++ openstack_plugin/resources/network/security_group.py
@@ -28,7 +28,8 @@ from openstack_plugin.constants import (RESOURCE_ID,
 from openstack_plugin.utils import (reset_dict_empty_keys,
                                     validate_resource_quota,
                                     add_resource_list_to_runtime_properties,
-                                    validate_ip_or_range_syntax)
+                                    validate_ip_or_range_syntax,
+                                    cleanup_runtime_properties)
 
 
 def security_group_creation_validation(openstack_resource):
@@ -98,14 +99,19 @@ def configure(openstack_resource, security_group_rules=None):
 
 
 @with_compat_node
-@with_openstack_resource(OpenstackSecurityGroup,
-                         ignore_unexisted_resource=True)
+@with_openstack_resource(OpenstackSecurityGroup)
 def delete(openstack_resource):
     """
     Delete current openstack security group instance
     :param openstack_resource: instance of openstack security group  resource
     """
+    if not ctx.instance.runtime_properties.get(RESOURCE_ID):
+        ctx.logger.info('SecurityGroup is already uninitialized.')
+        return
     openstack_resource.delete()
+    cleanup_runtime_properties(ctx, [
+        RESOURCE_ID
+    ])
 
 
 @with_compat_node
